use uint64_t

+++ modules/runtime/include/type/type.hpp
@@ -23,24 +23,24 @@ struct STypeRegistry
 };
 RUNTIME_API STypeRegistry* GetTypeRegistry();
 
-RUNTIME_API size_t Hash(bool value, size_t base);
-RUNTIME_API size_t Hash(int32_t value, size_t base);
-RUNTIME_API size_t Hash(int64_t value, size_t base);
-RUNTIME_API size_t Hash(uint32_t value, size_t base);
-RUNTIME_API size_t Hash(uint64_t value, size_t base);
-RUNTIME_API size_t Hash(float value, size_t base);
-RUNTIME_API size_t Hash(const skr_float2_t& value, size_t base);
-RUNTIME_API size_t Hash(const skr_float3_t& value, size_t base);
-RUNTIME_API size_t Hash(const skr_float4_t& value, size_t base);
-RUNTIME_API size_t Hash(const skr_float4x4_t& value, size_t base);
-RUNTIME_API size_t Hash(const skr_quaternion_t& value, size_t base);
-RUNTIME_API size_t Hash(const skr_rotator_t& value, size_t base);
-RUNTIME_API size_t Hash(double value, size_t base);
-RUNTIME_API size_t Hash(const skr_guid_t& value, size_t base);
-RUNTIME_API size_t Hash(const skr_md5_t& value, size_t base);
-RUNTIME_API size_t Hash(const skr_resource_handle_t& value, size_t base);
-RUNTIME_API size_t Hash(const skr::string& value, size_t base);
-RUNTIME_API size_t Hash(const skr::string_view& value, size_t base);
+RUNTIME_API uint64_t Hash(bool value, uint64_t base);
+RUNTIME_API uint64_t Hash(int32_t value, uint64_t base);
+RUNTIME_API uint64_t Hash(int64_t value, uint64_t base);
+RUNTIME_API uint64_t Hash(uint32_t value, uint64_t base);
+RUNTIME_API uint64_t Hash(uint64_t value, uint64_t base);
+RUNTIME_API uint64_t Hash(float value, uint64_t base);
+RUNTIME_API uint64_t Hash(const skr_float2_t& value, uint64_t base);
+RUNTIME_API uint64_t Hash(const skr_float3_t& value, uint64_t base);
+RUNTIME_API uint64_t Hash(const skr_float4_t& value, uint64_t base);
+RUNTIME_API uint64_t Hash(const skr_float4x4_t& value, uint64_t base);
+RUNTIME_API uint64_t Hash(const skr_quaternion_t& value, uint64_t base);
+RUNTIME_API uint64_t Hash(const skr_rotator_t& value, uint64_t base);
+RUNTIME_API uint64_t Hash(double value, uint64_t base);
+RUNTIME_API uint64_t Hash(const skr_guid_t& value, uint64_t base);
+RUNTIME_API uint64_t Hash(const skr_md5_t& value, uint64_t base);
+RUNTIME_API uint64_t Hash(const skr_resource_handle_t& value, uint64_t base);
+RUNTIME_API uint64_t Hash(const skr::string& value, uint64_t base);
+RUNTIME_API uint64_t Hash(const skr::string_view& value, uint64_t base);
 
 template <class T>
 auto GetCopyCtor();
@@ -59,8 +59,8 @@ struct RUNTIME_API skr_type_t {
     // clone type self
     skr_type_t* _Clone() const;
 
-    size_t Size() const;
-    size_t Align() const;
+    uint64_t Size() const;
+    uint64_t Align() const;
     skr_guid_t Id() const;
     const char8_t* Name() const;
     bool Same(const skr_type_t* srcType) const;
@@ -68,10 +68,10 @@ struct RUNTIME_API skr_type_t {
     void Convert(void* dst, const void* src, const skr_type_t* srcType, skr::type::ValueSerializePolicy* policy = nullptr) const;
     skr::string ToString(const void* dst, skr::type::ValueSerializePolicy* policy = nullptr) const;
     void FromString(void* dst, skr::string_view str, skr::type::ValueSerializePolicy* policy = nullptr) const;
-    size_t Hash(const void* dst, size_t base) const;
+    uint64_t Hash(const void* dst, uint64_t base) const;
     // lifetime operator
     void Destruct(void* dst) const;
-    void Construct(void* dst, skr::type::Value* args, size_t nargs) const;
+    void Construct(void* dst, skr::type::Value* args, uint64_t nargs) const;
     void* Malloc() const;
     void Free(void*) const;
     // copy construct
@@ -89,7 +89,7 @@ struct RUNTIME_API skr_type_t {
 struct RUNTIME_API skr_field_t {
     const char8_t* name SKR_IF_CPP(= nullptr);
     const skr_type_t* type SKR_IF_CPP(= nullptr);
-    size_t offset SKR_IF_CPP(= 0);
+    uint64_t offset SKR_IF_CPP(= 0);
 };
 
 struct SKR_ALIGNAS(16) RUNTIME_API skr_value_t {
@@ -100,7 +100,7 @@ struct SKR_ALIGNAS(16) RUNTIME_API skr_value_t {
         uint8_t _smallObj[24];
     };
 #ifdef __cplusplus
-    static constexpr size_t smallSize = sizeof(_smallObj);
+    static constexpr uint64_t smallSize = sizeof(_smallObj);
 
     skr_value_t() = default;
     skr_value_t(const skr_value_ref_t& ref);
@@ -136,7 +136,7 @@ struct SKR_ALIGNAS(16) RUNTIME_API skr_value_t {
     void* Ptr();
     const void* Ptr() const;
 
-    size_t Hash() const;
+    uint64_t Hash() const;
     skr::string ToString() const;
 
     void Reset();
@@ -188,7 +188,7 @@ struct RUNTIME_API skr_value_ref_t {
     template <class T>
     T Convert();
 
-    size_t Hash() const;
+    uint64_t Hash() const;
     skr::string ToString() const;
     void Reset();
 #endif
@@ -198,7 +198,7 @@ struct RUNTIME_API skr_method_t {
     const char* name SKR_IF_CPP(= nullptr);
     const skr_type_t* retType SKR_IF_CPP(= nullptr);
     const skr_field_t* parameters SKR_IF_CPP(= nullptr);
-    skr_value_t (*execute)(void* self, skr_value_ref_t* args, size_t nargs) SKR_IF_CPP(= nullptr);
+    skr_value_t (*execute)(void* self, skr_value_ref_t* args, uint64_t nargs) SKR_IF_CPP(= nullptr);
 };
 
 #ifdef __cplusplus
@@ -338,10 +338,10 @@ struct RUNTIME_API StringViewType : skr_type_t {
 // T[]
 struct RUNTIME_API ArrayType : skr_type_t {
     const struct skr_type_t* elementType;
-    size_t num;
-    size_t size;
+    uint64_t num;
+    uint64_t size;
     skr::string name;
-    ArrayType(const struct skr_type_t* elementType, size_t num, size_t size)
+    ArrayType(const struct skr_type_t* elementType, uint64_t num, uint64_t size)
         : skr_type_t{ SKR_TYPE_CATEGORY_ARR }
         , elementType(elementType)
         , num(num)
@@ -352,10 +352,10 @@ struct RUNTIME_API ArrayType : skr_type_t {
 // Object
 struct ObjectMethodTable {
     void (*dtor)(void* self);
-    void (*ctor)(void* self, Value* param, size_t nparam);
+    void (*ctor)(void* self, Value* param, uint64_t nparam);
     void (*copy)(void* self, const void* other);
     void (*move)(void* self, void* other);
-    size_t (*Hash)(const void* self, size_t base);
+    uint64_t (*Hash)(const void* self, uint64_t base);
     int (*Serialize)(const void* self, skr_binary_writer_t* writer);
     int (*Deserialize)(void* self, skr_binary_reader_t* reader);
     void (*SerializeText)(const void*, skr_json_writer_t* writer);
@@ -373,8 +373,8 @@ struct RUNTIME_API ArrayViewType : skr_type_t {
 };
 // struct/class T
 struct RUNTIME_API RecordType : skr_type_t {
-    size_t size = 0;
-    size_t align = 0;
+    uint64_t size = 0;
+    uint64_t align = 0;
     skr_guid_t guid = {};
     bool object = false; // true if inherits from SInterface
     const skr::string_view name = u8"";
@@ -385,7 +385,7 @@ struct RUNTIME_API RecordType : skr_type_t {
     bool IsBaseOf(const RecordType& other) const;
     RecordType() = default;
     RecordType(RecordType&&) = default;
-    RecordType(size_t size, size_t align, skr::string_view name, skr_guid_t guid, bool object, const RecordType* base, ObjectMethodTable nativeMethods,
+    RecordType(uint64_t size, uint64_t align, skr::string_view name, skr_guid_t guid, bool object, const RecordType* base, ObjectMethodTable nativeMethods,
     const skr::span<struct skr_field_t> fields, const skr::span<struct skr_method_t> methods)
         : skr_type_t{ SKR_TYPE_CATEGORY_OBJ }
         , size(size)
@@ -499,8 +499,8 @@ struct type_of<T&> {
     }
 };
 
-RUNTIME_API const skr_type_t* make_array_type(const skr_type_t* type, size_t num, size_t size);
-template <class T, size_t N>
+RUNTIME_API const skr_type_t* make_array_type(const skr_type_t* type, uint64_t num, uint64_t size);
+template <class T, uint64_t N>
 struct type_of<T[N]> {
     static const skr_type_t* get()
     {
@@ -509,7 +509,7 @@ struct type_of<T[N]> {
 };
 
 RUNTIME_API const skr_type_t* make_array_view_type(const skr_type_t* type);
-template <class T, size_t size>
+template <class T, uint64_t size>
 struct type_of<skr::span<T, size>> {
     static const skr_type_t* get()
     {
